chore: clean-up code for breaking up roadmap.ts
