+++ pages/roadmap/[...slug].tsx
@@ -13,8 +13,6 @@ import { ViewMode } from '../../lib/enums';
 import { setViewMode } from '../../hooks/useViewMode';
 import { DateGranularityState } from '../../lib/enums';
 import { setDateGranularity } from '../../hooks/useDateGranularity';
-import { useAsync } from '../../hooks/useAsync';
-import { paramsFromUrl } from '../../lib/paramsFromUrl';
 
 export async function getServerSideProps(context): Promise<ServerSidePropsResult> {
   const [hostname, owner, repo, issues_placeholder, issue_number] = context.query.slug;
@@ -36,7 +34,7 @@ export async function getServerSideProps(context): Promise<ServerSidePropsResult
       new URL(`${API_URL}?owner=${owner}&repo=${repo}&issue_number=${issue_number}&filter_group=${filter_group}`),
     );
     const response: RoadmapApiResponse = await res.json();
-    // console.log(`(response as RoadmapApiResponseSuccess)?.pendingChildren: `, (response as RoadmapApiResponseSuccess)?.pendingChildren);
+
     return {
       props: {
         ...serverSideProps,
@@ -60,56 +58,6 @@ export async function getServerSideProps(context): Promise<ServerSidePropsResult
 let done = false;
 export default function RoadmapPage(props: InferGetServerSidePropsType<typeof getServerSideProps>) {
   const { issueData, error, errors, isLocal, mode, dateGranularity, pendingChildren } = props;
-  const pendingChild = pendingChildren[0]
-  const {execute, status, value, error: asyncError} = useAsync(async () => {
-  //   for await (const pendingChild of pendingChildren) {
-    if (done) {
-      return;
-    }
-      done = true
-      console.log(`pendingChild: `, pendingChild);
-      const { issue_number, owner, repo } = paramsFromUrl(pendingChild.html_url)
-      const params = new URLSearchParams()
-      params.append('issue_number', issue_number);
-      params.append('repo', repo);
-      params.append('owner', owner);
-      params.append('group', pendingChild.group);
-      const url = new URL(`${window.location.origin}/api/pendingChild?${params}`)
-      console.log(`url: `, url);
-      try {
-        const res = await fetch(url);
-        const response: RoadmapApiResponse = await res.json();
-        console.log(`response: `, response);
-      } catch (err) {
-        console.error('error getting pending child', err);
-        // break;
-        throw err
-      }
-
-  //   }
-  }, false);
-  // console.log(`asyncError: `, asyncError);
-  // console.log(`value: `, value);
-  // console.log(`status: `, status);
-  // console.log(`execute: `, execute);
-
-  useEffect(() => {
-    switch(status) {
-      case 'idle':
-        execute();
-        break;
-      case 'error':
-        console.log('error', asyncError);
-        break;
-      case 'pending':
-        break;
-      case 'success':
-        console.log('success', value);
-
-        break;
-    }
-  }, [status, value, execute, asyncError])
-
 
   useEffect(() => {
     setDateGranularity(dateGranularity);
